Fix "hWaitForInput: invalid argument" errors
When we use fdToHandle', GHC is careful not to enabled non-blocking mode
on the file descriptor because it might affect other users of the FD.
But when the FD does not have non-blocking mode enabled, reading from it
uses select(), which returns EINVAL if the FD is >= 1024.  This is
really a bug and we should fix it in GHC.  But for hinotify, we can work
around it by allowing the IO library to enable non-blocking on our FD.

The API that GHC provides here isn't great, but it hasn't changed at all
for several years.